pyproject.toml
[project]
name = "psstlabs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=0.3.22",
    "openai>=1.68.2",
    "python-dotenv>=1.0.1",
    "simple-salesforce>=1.12.6",
    "tenacity>=9.0.0",
    "pandas>=2.2.2",
    "numpy>=2.0.1",
    "sacrebleu>=2.4.3",
    "rouge>=1.0.1",
    "google-genai>=0.2.0",
]
